refactor: Token domain repackage
