+++ src/main/java/com/playkuround/playkuroundserver/domain/auth/token/api/request/TokenReissueRequest.java
@@ -1,4 +1,4 @@
-package com.playkuround.playkuroundserver.domain.auth.token.dto.request;
+package com.playkuround.playkuroundserver.domain.auth.token.api.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
